% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{aggregate_signals}
\alias{aggregate_signals}
\title{Aggregate list of \code{covidcast_signal} objects into one data frame}
\usage{
aggregate_signals(x, format = c("wide", "long"))
}
\arguments{
\item{x}{List of \code{covidcast_signal} data frames.}

\item{format}{One of either "wide" or "long". The default is "wide".}
}
\value{
Data frame of aggregated signals in "wide" or "long" form, depending
on \code{format}. Important: in "wide" aggregation, only the latest issue from
each data frame is retained, and the columns \code{data_source}, \code{signal},
\code{issue}, \code{lag}, \code{stderr}, \code{sample_size} are all dropped from the returned
data frame.

For convenience, the function \code{aggregate_signals()} can operate on a list
of \code{covidcast_signal} data frames with time-shifted values (so one can
first call \code{append_shifts()}, and then \code{aggregate_signals()}). When there
are time-shifted values, and \code{format} is "tall", then a \code{dt} column gets
appended to the output data frame.
}
\description{
Aggregates a list of \code{covidcast_signal} objects into one data frame, in
either "wide" or "long" format. (In "wide" aggregation, only the latest issue
from each data frame is retained, and several columns are dropped.)
}
