% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{county_census}
\alias{county_census}
\title{County populations.}
\format{
An object of class \code{data.frame} with 3193 rows and 165 columns.
}
\source{
\url{https://www2.census.gov/programs-surveys/popest/datasets/2010-2019/counties/totals/co-est2019-alldata.csv}
}
\usage{
county_census
}
\description{
Data set on county populations, from the 2019 US Census.
}
\keyword{datasets}
