% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{county_fips}
\alias{county_fips}
\title{County FIPS codes.}
\format{
An object of class \code{data.frame} with 3193 rows and 3 columns.
}
\usage{
county_fips
}
\description{
Data set on county FIPS codes, extracted from the \code{county_pop} data frame.
}
\keyword{datasets}
