% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{county_geo}
\alias{county_geo}
\title{County latitudes and longitudes}
\format{
Data frame with 3142 rows, each representing one county. Columns:

\describe{
\item{x}{Longitude of county centroid.}
\item{y}{Latitude of county centroid.}
\item{fips}{Five-digit county FIPS code.}
\item{abbr}{Two-letter state abbreviation.}
\item{full}{State name.}
\item{county}{County name.}
}
}
\source{
\code{usmap}
}
\usage{
county_geo
}
\description{
Data set on latitudes and longitudes of county centroids, from the \code{usmap}
package.
}
\keyword{datasets}
