% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{covidcast_longer}
\alias{covidcast_longer}
\title{Pivot covidcast_signal_wide object to "long" format}
\usage{
covidcast_longer(x)
}
\description{
Pivots a \code{covidcast_signal_wide} object, the output of \code{aggregate_signals()}
with \code{format = "wide"}, into "long" format, i.e., as if \code{aggregate_signals()}
had been called in the first place with \code{format = "long"}.
}
