% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidcast.R
\name{covidcast_signal}
\alias{covidcast_signal}
\title{Produce a data frame for one signal.}
\usage{
covidcast_signal(
  data_source,
  signal,
  start_day = NULL,
  end_day = NULL,
  geo_type = c("county", "hrr", "msa", "dma", "state"),
  geo_values = "*"
)
}
\arguments{
\item{data_source}{String identifying the data source to query. See the
\href{https://cmu-delphi.github.io/delphi-epidata/api/covidcast_signals.html}{signal documentation}
for a list of available data sources.}

\item{signal}{String identifying the signal from that source to query. Again,
see the \href{https://cmu-delphi.github.io/delphi-epidata/api/covidcast_signals.html}{signal documentation}
for a list of available signals.}

\item{start_day}{Query data beginning on this date. Date object, or string in
the form YYYYMMDD. If \code{start_day} is \code{NULL}, defaults to first day data is
available for this signal.}

\item{end_day}{Query data up to this date, inclusive. Date object or string
in the form YYYYMMDD. If \code{end_day} is \code{NULL}, defaults to the most recent
day data is available for this signal.}

\item{geo_type}{The geography type for which to request this data, such as
\code{"county"} or \code{"state"}. Defaults to \code{"county"}. See the \href{https://cmu-delphi.github.io/delphi-epidata/api/covidcast_geography.html}{geographic coding documentation}
for details on which types are available.}

\item{geo_values}{Which geographies to return. The default, \code{"*"}, fetches
all geographies. To fetch specific geographies, specify their IDs as a
vector or list of strings. See the \href{https://cmu-delphi.github.io/delphi-epidata/api/covidcast_geography.html}{geographic coding documentation}
for details on how to specify these IDs.}
}
\value{
Data frame with matching data. Each row is one observation on one day
in one geographic location. Contains the following columns:

\item{geo_value}{identifies the location, such as a state name or county
FIPS code}
\item{time_value}{a \code{Date} object identifying the date of this observation}
\item{value}{the signal quantity requested. For example, in a query for the
\code{confirmed_cumulative_num} signal from the \code{usa-facts} source, this would
be the cumulative number of confirmed cases in the area, as of the
\code{time_value}.}
\item{stderr}{the value's standard error, if available}
\item{sample_size}{indicates the sample size available in that geography on
that day; sample size may not be available for all signals, due to privacy
or other constraints, in which case they will be \code{NA}.}
\item{direction}{uses a local linear fit to estimate whether the signal in
this region is currently increasing or decreasing (reported as -1 for
decreasing, 1 for increasing, and 0 for neither).}

Consult the signal documentation for more details on how values and
standard errors are calculated for specific signals.
}
\description{
Obtains data for selected date ranges for all geographic regions of the
United States. Available data sources and signals are documented in the
\href{https://cmu-delphi.github.io/delphi-epidata/api/covidcast_signals.html}{COVIDcast signal documentation}.
Most (but not all) data sources are available at the county level, but the
API can also return data aggregated to metropolitan statistical areas,
hospital referral regions, or states, as desired, by using the \code{geo_type}
argument. View \code{vignette("covidcastR")} for detailed example usage.
}
\details{
Downloading large amounts of data may be slow, so this function prints
messages for each day of data it downloads. To suppress these, use
\code{\link[base:message]{base::suppressMessages()}}, as in \code{suppressMessages(covidcast_signal("fb-survey", ...))}.
}
\examples{
\dontrun{
## fetch all counties from 2020-05-10 to the most recent available data:
covidcast_signal("fb-survey", "raw_cli", start_day = "20200510")
## fetch all counties on just 2020-05-10 and no other days
covidcast_signal("fb-survey", "raw_cli", start_day = "20200510",
                 end_day = "20200510")
## fetch all states on 2020-05-10, 2020-05-11, 2020-05-12
covidcast_signal("fb-survey", "raw_cli", start_day = "20200510",
                 end_day = "20200512", geo_type = "state")
## fetch all available data for just Pennsylvania and New Jersey
covidcast_signal("fb-survey", "raw_cli", geo_type = "state",
                 geo_values = c("pa", "nj"))
## fetch all available data in Pittsburgh metropolitan area (identified by
## CBSA ID)
covidcast_signal("fb-survey", "raw_cli", geo_type = "msa",
                 geo_values = "38300")
}
}
\references{
COVIDcast API documentation:
\url{https://cmu-delphi.github.io/delphi-epidata/api/covidcast.html}

Documentation of all COVIDcast sources and signals:
\url{https://cmu-delphi.github.io/delphi-epidata/api/covidcast_signals.html}

COVIDcast public map: \url{https://covidcast.cmu.edu/}
}
