% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidcast.R
\name{covidcast_signals}
\alias{covidcast_signals}
\title{Obtain multiple COVIDcast signals in one data frame}
\usage{
covidcast_signals(
  data_source,
  signal,
  start_day = NULL,
  end_day = NULL,
  geo_type = c("county", "hrr", "msa", "dma", "state"),
  geo_values = "*",
  as_of = NULL,
  issues = NULL,
  lag = NULL
)
}
\description{
This convenience function uses \code{covidcast_signal()} to obtain multiple
signals, potentially from multiple data sources.
}
\details{
The argument structure is just as in \code{covidcast_signal()}, except
the first four arguments \code{data_source}, \code{signal}, \code{start_day}, \code{end_day}
are permitted to be vectors. The returned object is a list of
\code{covidcast_signal} data frames, of length \code{N = max(length(data_source), length(signal))}. The first two arguments \code{data_source}, \code{signal} are
recycled appropriately, in the calls to \code{covidcast_signal()}; see example
below. The next two arguments \code{start_day}, \code{end_day}, unless \code{NULL}, must
be either length 1 or N.
}
\examples{
\dontrun{
## Fetch USAFacts confirmed cases and deaths over the same time period
covidcast_signals("usa-facts", signal=c("confirmed_incidence_num",
                                        "deaths_incidence_num"),
                   start_day = "2020-08-15", end_day = "2020-10-01")
}
}
\seealso{
\code{\link[=covidcast_signal]{covidcast_signal()}}
}
