% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{covidcast_wider}
\alias{covidcast_wider}
\title{Pivot covidcast_signal_long object to "wide" format}
\usage{
covidcast_wider(x)
}
\description{
Pivots a \code{covidcast_signal_long} object, the output of \code{aggregate_signals()}
with \code{format = "long"}, into "wide" format, i.e., as if \code{aggregate_signals()}
had been called in the first place with \code{format = "wide"}.
}
