% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{earliest_issue}
\alias{earliest_issue}
\title{Fetch only the earliest issue for each observation in a data frame}
\usage{
earliest_issue(df)
}
\arguments{
\item{df}{A \code{covidcast_signal} data frame}
}
\value{
The same \code{covidcast_signal} data frame, but with only the earliest
issue of every observation
}
\description{
Since \code{covidcast_signal()} can, with the right options, return multiple
issues for a single observation in a single geo, we may want only the most
recent for plotting, mapping, or other purposes.
}
\keyword{internal}
