% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{msa_cbsa}
\alias{msa_cbsa}
\title{Metro area CBSA codes.}
\format{
An object of class \code{data.frame} with 2797 rows and 2 columns.
}
\usage{
msa_cbsa
}
\description{
Data set on metropolitan area CBSA codes, extracted from the \code{msa_pop} data frame.
}
\keyword{datasets}
