% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{msa_pop}
\alias{msa_pop}
\title{Metro area populations.}
\format{
An object of class \code{data.frame} with 2797 rows and 97 columns.
}
\source{
\url{https://www2.census.gov/programs-surveys/popest/datasets/2010-2019/metro/totals/cbsa-est2019-alldata.csv}
}
\usage{
msa_pop
}
\description{
Data set metropolitan area populations, from the 2019 US Census.
}
\keyword{datasets}
