% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidcast.R
\name{plot.covidcast_signal}
\alias{plot.covidcast_signal}
\title{Plot function for covidcast_signal object}
\usage{
\method{plot}{covidcast_signal}(
  x,
  plot_type = c("choro", "bubble", "line"),
  time_values = NULL,
  include = c(),
  geo_values = NULL,
  range = NULL,
  choro_col = c("#FFFFCC", "#FD893C", "#800026"),
  alpha = 0.5,
  direction = FALSE,
  dir_col = c("#6F9CC6", "#F7F793", "#C56B59"),
  bubble_col = "purple",
  line_col = 1:6,
  lty = 1:5,
  title = NULL,
  choro_params = list(),
  bubble_params = list(),
  line_params = list()
)
}
\arguments{
\item{x}{The \code{covidcast_signal} object.}

\item{plot_type}{One of "choro", "bubble", "line" indicating whether to plot
a choropleth map, bubble map, or line (time series) graph, respectively.
The default is "choro".}

\item{time_values}{Vector of Date objects (or strings in the form YYYY-MM-DD)
specifying the days for plotting. For choropleth and bubble maps, only a
single day can be plotted at one time, and this is taken to be maximum date
in \code{time_values}. If \code{NULL}, the default, then the last date in \code{x} is used
for the maps, and (up to) the last 2 weeks in \code{x} is used for the time
series plot.}

\item{include}{Vector of state abbreviations (case insensitive, so "pa" and
"PA" both denote Pennsylvania) indicating which states to include in the
choropleth and bubble maps. Default is \code{c()}, which is interpreted to mean
all states.}

\item{geo_values}{Vector of \code{geo_values} to include in the time series
plot. If \code{NULL}, the default, then the first 6 \code{geo_values} as found in \code{x}
are used.}

\item{range}{Vector of two values: min and max, in this order, to use when
defining the color scale for choropleth maps and the size scale for bubble
maps. If \code{NULL}, the default, then the min and max are set to be the mean
+/- 3 standard deviations, where this mean and standard deviation are as
provided in the meta data for the given data source and signal.}

\item{choro_col}{Vector of colors, as specified in hex code, to use for the
choropleth color scale. Can be arbitrary in length. Default is similar to
that from covidcast.cmu.edu.}

\item{alpha}{Number between 0 and 1, indicating the transparency level to be
used in the maps. For choropleth maps, this determines the transparency
level for the mega counties. For bubble maps, this determines the
transparency level for the bubbles. Default is 0.5.}

\item{direction}{Should direction be visualized (instead of intensity) for
the maps? Default is \code{FALSE}.}

\item{dir_col}{Vector of colors, as specified in hex code, to use for the
direction color scale. Must be of length 3. Default is similar to that from
covidcast.cmu.edu.}

\item{bubble_col}{Bubble color for the bubble map. Default is "purple".}

\item{line_col}{Vector of colors for the time series plot. This will be
recycled as necessary. Default is \code{1:6}.}

\item{lty}{Vector of line types for the time series plot. This will be
recycled as necessary. Default is \code{1:5}.}

\item{title}{Title for the plot. If \code{NULL}, the default, then a simple title
is used based on the given data source, signal, and time values.}

\item{choro_params, bubble_params, line_params}{Additional parameter lists for
the different plot types, for further customization. See details below.}
}
\description{
Plot function for covidcast_signal object
}
