% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidcast.R
\name{print.covidcast_meta}
\alias{print.covidcast_meta}
\title{Print \code{covidcast_meta} object}
\usage{
\method{print}{covidcast_meta}(x, ...)
}
\arguments{
\item{x}{The \code{covidcast_meta} object.}

\item{...}{Additional arguments passed to \code{print.data.frame()} to print the
data.}
}
\description{
Prints a brief summary of the metadata, and then prints the underlying data
frame, for an object returned by \code{covidcast_meta()}.
}
