% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidcast.R
\name{print.covidcast_signal}
\alias{print.covidcast_signal}
\title{Print function for covidcast_signal object}
\usage{
\method{print}{covidcast_signal}(x)
}
\arguments{
\item{x}{The \code{covidcast_signal} object.}
}
\description{
Print function for covidcast_signal object
}
