% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{state_fips}
\alias{state_fips}
\title{State FIPS codes.}
\format{
An object of class \code{data.frame} with 53 rows and 2 columns.
}
\usage{
state_fips
}
\description{
Data set on state FIPS codes, extracted from the \code{state_pop} data frame.
}
\keyword{datasets}
