% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{state_geo}
\alias{state_geo}
\title{State latitudes and longitudes}
\format{
Data frame with 51 rows, each representing one state (including the
District of Columbia). Columns:

\describe{
\item{x}{Longitude of state centroid.}
\item{y}{Latitude of state centroid.}
\item{fips}{Five-digit county FIPS code.}
\item{abbr}{Two-letter state abbreviation.}
\item{full}{State name.}
}
}
\source{
\code{usmap}
}
\usage{
state_geo
}
\description{
Data set on latitudes and longitudes of state centroids, from the \code{usmap}
package.
}
\keyword{datasets}
