% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{state_pop}
\alias{state_pop}
\title{State populations.}
\format{
An object of class \code{data.frame} with 53 rows and 8 columns.

An object of class \code{data.frame} with 53 rows and 8 columns.
}
\source{
\url{https://www2.census.gov/programs-surveys/popest/datasets/2010-2019/state/detail/SCPRC-EST2019-18+POP-RES.csv}

\url{https://www2.census.gov/programs-surveys/popest/datasets/2010-2019/state/detail/SCPRC-EST2019-18+POP-RES.csv}
}
\usage{
state_pop

state_pop
}
\description{
Data set on state populations, from the 2019 US Census.

Data set on state populations, from the 2019 US Census.
}
\keyword{datasets}
