% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{append_shifts}
\alias{append_shifts}
\title{Append time-shifted values to a \code{covidcast_signal} data frame}
\usage{
append_shifts(x, dt)
}
\arguments{
\item{x}{The \code{covidcast_signal} data frame.}

\item{dt}{Vector of shifts to append for the values in \code{x}. Positive values
are taken to mean shifts forward in time, and negative values mean shifts
backward in time. For example, if \code{dt = 1}, then the values are shifted
forward 1 day in time (so, data on June 1 becomes data on June 2, and so
on). Default is 0 for both.}
}
\value{
A data frame one new column per value of \code{dt}. The new column names
reflect the values of \code{dt}, as in \code{value+1} and \code{value-1} for a time shift
forward 1 day and backward 1 day, respectively.
}
\description{
Appends time-shifted values, lagging or leading, to a \code{covidcast_signal} data
frame. (Only the latest issue from each data frame is preserved.)
}
