% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{county_geo}
\alias{county_geo}
\title{County latitudes and longitudes}
\format{
Data frame with 3331 rows, each representing one county. Columns:

\describe{
\item{COUNTYNAME}{Name of the county.}
\item{FIPS}{5-digit county FIPS code.}
\item{STATE}{Two-letter state abbreviation.}
\item{LON}{Longitude of county centroid.}
\item{LAT}{Latitude of county centroid.}
}
}
\source{
\url{https://www.weather.gov/gis/Counties}
}
\usage{
county_geo
}
\description{
Data set on latitudes and longitudes of county centroids, from the National
Weather Service.
}
\keyword{datasets}
