% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{county_geo}
\alias{county_geo}
\title{County latitudes and longitudes.}
\format{
An object of class \code{data.frame} with 3331 rows and 7 columns.
}
\source{
\url{https://www.weather.gov/gis/Counties}
}
\usage{
county_geo
}
\description{
Data set on latitudes and longitudes of county centroids, from the National
Weather Service.
}
\keyword{datasets}
