% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.R
\name{covidcast_cor}
\alias{covidcast_cor}
\title{Compute correlations between two \code{covidcast_signal} data frames}
\usage{
covidcast_cor(
  x,
  y,
  dt_x = 0,
  dt_y = 0,
  by = c("geo_value", "time_value"),
  use = "na.or.complete",
  method = c("spearman", "pearson", "kendall")
)
}
\arguments{
\item{x, y}{The \code{covidcast_signal} data frames to correlate.}

\item{dt_x, dt_y}{Time shifts to consider for \code{x} and \code{y}, respectively,
before computing correlations. These must both be nonnegative, and only one
can be positive. If \code{dt_x = 1}, for example, then data for \code{x} is shifted
forward 1 day in time (so, data on June 1 becomes data on June 2, and so
on). Default is 0 for both.}

\item{by}{If "geo_value", then correlations are computed for each geo
location, over all time. Each correlation is measured between two time
series at the same location. If "time_value", then correlations are
computed for each time, over all geo locations. Each correlation is
measured between all locations at one time. Default is "geo_value".}

\item{use, method}{Arguments to pass to \code{cor()}, with "na.or.complete" the
default for \code{use}, and "spearman" the default for \code{method} (different than
the defaults used by \code{cor()}).}
}
\value{
A data frame with first column \code{geo_value} or \code{time_value} (depending
on \code{by}), and second column \code{value}, which gives the correlation.
}
\description{
Computes correlations between two \code{covidcast_signal()} data frames, allowing
for slicing by geo location, or by time. (The latest issue from each data
frame is used when computing correlations.) See
\code{vignette("correlation-utils")} for examples.
}
