% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fips_to_abbr}
\alias{fips_to_abbr}
\title{Get state abbreviations from FIPS codes}
\usage{
fips_to_abbr(
  code,
  ignore.case = TRUE,
  perl = FALSE,
  fixed = FALSE,
  ties_method = c("first", "all")
)
}
\arguments{
\item{code}{Vector of FIPS codes to look up; codes can have either two digits
(as in "42") or five digits (as in "42000"), either is allowed.}

\item{ignore.case, perl, fixed}{Arguments to pass to \code{grep()}, with the same
defaults as in the latter function. Hence, by default, regular expressions
are used; to match against a fixed string (no regular expressions), set
\code{fixed = TRUE}.}

\item{ties_method}{If "first", then only the first match for each code is
returned. If "all", then all matches for each code are returned.}
}
\value{
A vector of state abbreviations if \code{ties_method} equals "first", and
a list of state abbreviations otherwise.
}
\description{
Look up state abbreviations by FIPS codes (including District of Columbia and
Puerto Rico); this function is based on \code{grep()}, and hence allows for
regular expressions.
}
\examples{
fips_to_abbr("42000")
fips_to_abbr(c("42", "72", "11"))

# Note that fips_to_name() works for state names too:
fips_to_name("42000")

}
\seealso{
\code{\link[=abbr_to_fips]{abbr_to_fips()}}
}
