% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidcast.R
\name{plot.covidcast_signal}
\alias{plot.covidcast_signal}
\title{Plot \code{covidcast_signal} objects}
\usage{
\method{plot}{covidcast_signal}(
  x,
  plot_type = c("choro", "bubble", "line"),
  time_value = NULL,
  include = c(),
  range = NULL,
  choro_col = c("#FFFFCC", "#FD893C", "#800026"),
  alpha = 0.5,
  direction = FALSE,
  dir_col = c("#6F9CC6", "#F3EE9E", "#C56B59"),
  bubble_col = "purple",
  num_bins = 8,
  line_col = 1:6,
  line_type = rep(1:6, each = length(line_col)),
  title = NULL,
  choro_params = list(),
  bubble_params = list(),
  line_params = list(),
  ...
)
}
\arguments{
\item{x}{The \code{covidcast_signal} object to map or plot. If the object
contains multiple issues of the same observation, only the most recent
issue is mapped or plotted.}

\item{plot_type}{One of "choro", "bubble", "line" indicating whether to plot
a choropleth map, bubble map, or line (time series) graph, respectively.
The default is "choro".}

\item{time_value}{Date object (or string in the form \code{"YYYY-MM-DD"})
specifying the day to map, for choropleth and bubble maps. If \code{NULL}, the
default, then the last date in \code{x} is used for the maps. Time series plots
always include all available time values in \code{x}.}

\item{include}{Vector of state abbreviations (case insensitive, so "pa" and
"PA" both denote Pennsylvania) indicating which states to include in the
choropleth and bubble maps. Default is \code{c()}, which is interpreted to mean
all states.}

\item{range}{Vector of two values: min and max, in this order, to use when
defining the color scale for choropleth maps and the size scale for bubble
maps, or the range of the y-axis for the time series plot. If \code{NULL}, the
default, then the min and max are set to be the mean +/- 3 standard
deviations, where this mean and standard deviation are as provided in the
meta data for the given data source and signal.}

\item{choro_col}{Vector of colors, as specified in hex code, to use for the
choropleth color scale. Can be arbitrary in length. Default is similar to
that from covidcast.cmu.edu.}

\item{alpha}{Number between 0 and 1, indicating the transparency level to be
used in the maps. For choropleth maps, this determines the transparency
level for the mega counties. For bubble maps, this determines the
transparency level for the bubbles. Default is 0.5.}

\item{direction}{Should direction be visualized (instead of intensity) for
the choropleth map? Default is \code{FALSE}.}

\item{dir_col}{Vector of colors, as specified in hex code, to use for the
direction color scale. Must be of length 3. Default is similar to that from
covidcast.cmu.edu.}

\item{bubble_col}{Bubble color for the bubble map. Default is "purple".}

\item{num_bins}{Number of bins for determining the bubble sizes for the
bubble map. Default is 6. These bins are evenly-spaced in between the min
and max as specified through the \code{range} parameter. Each bin is assigned
the same bubble size. Also, values of zero special: it has its own separate
(small) bin, and values mapped to the zero bin are not drawn.}

\item{line_col}{Vector of colors for the time series plot. This will be
recycled as necessary. Default is \code{1:6}.}

\item{line_type}{Vector of line types for the time series plot. This will be
recycled as necessary. Default is \code{rep(1:6, each = length(col))}.}

\item{title}{Title for the plot. If \code{NULL}, the default, then a simple title
is used based on the given data source, signal, and time values.}

\item{choro_params, bubble_params, line_params}{Additional parameter lists for
the different plot types, for further customization. See details below.}

\item{...}{Additional arguments, for compatibility with \code{plot()}. Currently
unused.}
}
\description{
Several plot types are provided, including choropleth plots (maps), bubble
plots, and time series plots showing the change of signals over time.
}
\details{
TODO explain all these advanced parameters.

For both choropleth and bubble maps:
\itemize{
\item{\code{subtitle}}
\item{\code{missing_col}}
\item{\code{border_col}}
\item{\code{border_size}}
\item{\code{legend_height}}
\item{\code{legend_width}}
\item{\code{breaks}}
}

For choropleth maps only:
\itemize{
\item{\code{legend_n}}
}

For bubble maps only:
\itemize{
\item{\code{remove_zero}}
\item{\code{min_size}}
\item{\code{max_size}}
}

For line graphs:
\itemize{
\item{\code{xlab}}
\item{\code{ylab}}
}
}
