% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{state_census}
\alias{state_census}
\title{State population data}
\format{
Data frame with 53 rows (including one for the United States as a
whole, plus the District of Columbia and the Puerto Rico Commonwealth).
Important columns:

\describe{
\item{NAME}{Name of the state.}
\item{POPESTIMATE2019}{Estimate of the state's resident population in 2019.}
}
}
\source{
United States Census Bureau, at
\url{https://www2.census.gov/programs-surveys/popest/datasets/2010-2019/state/detail/SCPRC-EST2019-18+POP-RES.csv}
}
\usage{
state_census
}
\description{
Data set on state populations, from the 2019 US Census.
}
\seealso{
\code{\link[=abbr_to_name]{abbr_to_name()}}, \code{\link[=name_to_abbr]{name_to_abbr()}}, \code{\link[=abbr_to_fips]{abbr_to_fips()}}, \code{\link[=fips_to_abbr]{fips_to_abbr()}}
}
\keyword{datasets}
