% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{state_geo}
\alias{state_geo}
\title{State latitudes and longitudes}
\format{
Data frame with 52 rows, each representing one state (including
Puerto Rico and the District of Columbia). Columns:

\describe{
\item{STATE}{Two-letter state abbreviation.}
\item{LAT}{Latitude of state centroid.}
\item{LON}{Longitude of state centroid.}
\item{NAME}{Name of state.}
}
}
\source{
\url{https://developers.google.com/public-data/docs/canonical/states_csv}
}
\usage{
state_geo
}
\description{
Data set on latitudes and longitudes of state centroids, from Google's DSPL.
}
\keyword{datasets}
