% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{state_geo}
\alias{state_geo}
\title{State latitudes and longitudes.}
\format{
An object of class \code{data.frame} with 52 rows and 4 columns.
}
\source{
\url{https://developers.google.com/public-data/docs/canonical/states_csv}
}
\usage{
state_geo
}
\description{
Data set on latitudes and longitudes of state centroids, from Google's DSPL.
}
\keyword{datasets}
