% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidcast.R
\name{summary.covidcast_meta}
\alias{summary.covidcast_meta}
\title{Summarize \code{covidcast_meta} object}
\usage{
\method{summary}{covidcast_meta}(object, ...)
}
\arguments{
\item{object}{The \code{covidcast_meta} object.}

\item{...}{Additional arguments, for compatibility with \code{summary()}.
Currently unused.}
}
\description{
Prints a tabular summary of the object returned by \code{covidcast_meta()}.
}
