% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidcast.R
\name{summary.covidcast_signal}
\alias{summary.covidcast_signal}
\title{Summarize \code{covidcast_signal} object}
\usage{
\method{summary}{covidcast_signal}(object, ...)
}
\arguments{
\item{object}{The \code{covidcast_signal} object.}

\item{...}{Additional arguments, for compatibility with \code{summary()}.
Currently unused.}
}
\description{
Prints a variety of summary statistics about the underlying data, such as
median values, the date range included, sample sizes, and so on, for an
object returned by \code{covidcast_signal()}.
}
